requirements-dev.txt
# 测试框架
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.2.0

# 代码质量
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# 开发工具
pre-commit==3.5.0

# 测试数据工厂
factory-boy==3.3.0

# HTTP测试
requests==2.31.0